Split compound format tests into separate base class
