+++ Elements.Quantity.Tests/Quantities/Basic/AccelerationTests.cs
@@ -1,5 +1,4 @@
-﻿using System.Collections.Generic;
-using System.Diagnostics.CodeAnalysis;
+﻿using System.Diagnostics.CodeAnalysis;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace Elements.Quantity.Test.Quantities.Basic;
@@ -13,7 +12,4 @@ public sealed class AccelerationTests : BaseQuantityTests<AccelerationTests, Acc
     [
         new(Acceleration.MetersPerSecondPerSecond, "{0} m/s^2", "1 meter per second squared", "{0} meters per second squared", "Quantity.Unit.Acceleration.MetersPerSecondSquared")
     ];
-
-    /// <inheritdoc/>
-    public static IEnumerable<QuantityFormatTestData<Acceleration>>? CompoundFormatInfoDataTuples => null;
 }
