+++ Elements.Quantity.Tests/Quantities/Basic/RatioTests.cs
@@ -1,5 +1,4 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 
 namespace Elements.Quantity.Test.Quantities.Basic;
@@ -14,7 +13,4 @@ public class RatioTests : BaseQuantityTests<RatioTests, Ratio>, IQuantityTestDat
         new(Ratio.RatioValue, "{0}", "1", "{0}", "Quantity.Unit.Ratio"),
         new(Ratio.Percent, "{0} %", "1 percent", "{0} percent", "Quantity.Unit.Ratio.Percent")
     ];
-
-    /// <inheritdoc/>
-    public static IEnumerable<QuantityFormatTestData<Ratio>>? CompoundFormatInfoDataTuples => null;
 }
