+++ Elements.Quantity.Tests/Quantities/Basic/VelocityTests.cs
@@ -1,5 +1,4 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 
 
@@ -18,7 +17,4 @@ public class VelocityTests : BaseQuantityTests<VelocityTests, Velocity>, IQuanti
         new(Velocity.FeetPerSecond, "{0} ft/s", "1 foot per second", "{0} feet per second", "Quantity.Unit.Velocity.FeetPerSecond"),
         new(Velocity.Knots, "{0} kn", "1 knot", "{0} knots", "Quantity.Unit.Velocity.Knots")
     ];
-
-    /// <inheritdoc/>
-    public static IEnumerable<QuantityFormatTestData<Velocity>>? CompoundFormatInfoDataTuples => null;
 }
