+++ Elements.Quantity.Tests/Quantities/IQuantityTestData.cs
@@ -1,5 +1,3 @@
-using System.Collections.Generic;
-
 namespace Elements.Quantity.Test.Quantities;
 
 public interface IQuantityTestData<TQuantity>
@@ -14,14 +12,4 @@ where TQuantity : unmanaged, IQuantity<TQuantity>
     /// This property is intended for use in unit tests.
     /// </remarks>
     static abstract QuantityTestData<TQuantity>[] TestDataTuples { get; }
-
-    /// <summary>
-    /// An array of test data tuples representing different combinations of <see cref="CompoundFormatInfo{TQuantity}"/> and
-    /// <typeparamref name="TQuantity"/> with their expected text representation.
-    /// Each element in the array contains the format, input base value, and expected string.
-    /// </summary>
-    /// <remarks>
-    /// This property is intended for use in unit tests. Return <code>null</code> if there are no test cases.
-    /// </remarks>
-    static abstract IEnumerable<QuantityFormatTestData<TQuantity>>? CompoundFormatInfoDataTuples { get; }
 }
